refactor: use simple safe and unsafe link buffer defination
